refactor; fixed lint err
